fix output of bold_seq function
In the output of bold_seq(), id and gene were both the id, there was a mistake in the indexes of split_fasta() on line 11 of zzz.R .

Some code in split_fasta() seemed to be redundant and was removed/replaced. I tried optimizing it for both time and memory usage.