update doc, response parsing, output attributes
Since most responses will have consistent numbers of fields and names, it's faster to get all fields and put them directly in a matrix by rows.

Added attributes to the output: 'errors' so errors can be found more easily and retried if needed and 'sequence' to make sure the sequences are not lost whether or not the user names the given vector/list of sequences.