update bold_tax_id
extracted the 'dataTypes' arg check into a function for readability.
Also replaced the argument check errors by warning to not affect existing workflows (since it didn't use to throw errors) and made it so, if any 'dataTypes' isn't valid, they are skipped (so not put in the url for the request) and if none are, it automatically returns NA without doing the API request at all.
Readded test for it.