Update utils.R
update to assert to be able to check params needing not to be empty or needing to be of length one.
update to setread to be able to add more options if needed.
new function to fix taxon names with single quotes or ending with a parenthesis or dot. see issues #84 & #85